pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kakeru"
version = "0.1.0"
description = "X (Twitter) Automation System"
requires-python = ">=3.12"
dependencies = [
    "selenium>=4.0.0",
    "webdriver-manager>=3.8.0",
    "python-dotenv>=0.19.0",
    "openai>=1.0.0",
    "tweepy>=4.0.0",
    "pyyaml>=6.0.0",
    "schedule>=1.1.0",
    "undetected-chromedriver>=3.5.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[tool.setuptools]
packages = ["kakeru", "bot"]
